chore: 租户相关内容从opne目录下剥离
